+++ saleor/app/installation_utils.py
@@ -55,7 +55,7 @@ def install_app(app_installation: AppInstallation, activate: bool = False):
 
     manifest_data["permissions"] = get_permission_names(assigned_permissions)
 
-    clean_manifest_data(manifest_data)
+    clean_manifest_data(manifest_data, raise_for_saleor_version=True)
 
     app = App.objects.create(
         name=app_installation.app_name,