+++ saleor/app/manifest_validations.py
@@ -1,11 +1,13 @@
 import logging
 from collections import defaultdict
-from typing import Dict, Iterable, List
+from typing import Dict, Iterable, List, Optional
 
 from django.core.exceptions import ValidationError
 from django.db.models import Value
 from django.db.models.functions import Concat
+from semantic_version import NpmSpec, Version
 
+from .. import __version__
 from ..graphql.core.utils import str_to_enum
 from ..graphql.webhook.subscription_query import SubscriptionQuery
 from ..permission.enums import (
@@ -95,7 +97,7 @@ def clean_permissions(
     return [p for p in saleor_permissions if p.codename in permissions]
 
 
-def clean_manifest_data(manifest_data):
+def clean_manifest_data(manifest_data, raise_for_saleor_version=False):
     errors: T_ERRORS = defaultdict(list)
 
     validate_required_fields(manifest_data, errors)
@@ -109,6 +111,13 @@ def clean_manifest_data(manifest_data):
                 code=AppErrorCode.INVALID_URL_FORMAT.value,
             )
         )
+
+    try:
+        manifest_data["requiredSaleorVersion"] = clean_required_saleor_version(
+            manifest_data.get("requiredSaleorVersion"), raise_for_saleor_version
+        )
+    except ValidationError as e:
+        errors["requiredSaleorVersion"].append(e)
 
     saleor_permissions = get_permissions().annotate(
         formated_codename=Concat("content_type__app_label", Value("."), "codename")
@@ -291,3 +300,27 @@ def validate_required_fields(manifest_data, errors):
                     code=AppErrorCode.REQUIRED.value,
                 )
             )
+
+
+def parse_version(version_str: str) -> Version:
+    return Version(version_str)
+
+
+def clean_required_saleor_version(
+    required_version,
+    raise_for_saleor_version: bool,
+    saleor_version=__version__,
+) -> Optional[Dict]:
+    if not required_version:
+        return None
+    try:
+        spec = NpmSpec(required_version)
+    except Exception:
+        msg = "Incorrect value for required Saleor version."
+        raise ValidationError(msg, code=AppErrorCode.INVALID.value)
+    version = parse_version(saleor_version)
+    satisfied = spec.match(version)
+    if raise_for_saleor_version and not satisfied:
+        msg = f"Saleor version {saleor_version} is not supported by the app."
+        raise ValidationError(msg, code=AppErrorCode.UNSUPPORTED_SALEOR_VERSION.value)
+    return {"constraint": required_version, "satisfied": satisfied}